// src/app/models/schedule.ts
export interface Schedule {
  id: number;
  title: string;
  startDate: string;
  finishDate: string;
  percentageComplete: string;
  scheduleStatus: string;
  observation: string;
  hrResult: string;
  supervisorResult: string;
  hrObservation: string;
  supervisorObservation: string;
  scheduleComplete: string;
}